Update README.md
This project work for in depth Data Analysis of a US Super Store Sales Data and helps for drawing meaningful insights which can help the Management Team for taking crucial decisions in order to expand there buisness and make an impactable market presence around the globe.

Please find attached Report file to understand the insights and working of the Dashboard. Things like ETL process, cleaning methods are also mentioned in it.